Pathfinding 1.0
The pathfinding sequence can now run thorugh without crashing
Still have not run to completion possible backtrack malfunction
Takes VERY LONG
Finish sequence is not complete

+++ TrafficSim/Roads/IRoad.cs
@@ -4,7 +4,7 @@
 
     {
         int NumberOfLanes{get; set; }
-        float TimeToTraverse {get; set;}
+        int TimeToTraverse {get; set;}
         double Cost { get; set; }
         ushort DirectionAOccupancy { get; set; }
         ushort DirectionBOccupancy { get; set; }
